% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Calculation of the bma object}
\usage{
bma(for_bma, df, round = 4, EMS = NULL, dilution = 0, dil.Par = 0.5)
}
\arguments{
\item{for_bma}{List with model space and likelihood table (the result of bma_prep function).}

\item{df}{Data frame with data for the SEM analysis.}

\item{round}{Parameter indicating the decimal place to which number in the BMA tables and prior and posterior model sizes should be rounded (default round = 4)}

\item{EMS}{Expected model size for model binomial and binomial-beta model prior}

\item{dilution}{Binary parameter: 0 - NO application of a dilution prior; 1 - application of a dilution prior (George 2010).}

\item{dil.Par}{Parameter associated with dilution prior - the exponent of the determinant (George 2010). Used only if parameter dilution = 1.}
}
\value{
A list with bma objects: \cr
\enumerate{
\item uniform_table - table with the results under binomial model prior \cr
\item random_table - table with the results under binomial-beta model prior \cr
\item reg_names - vector with names of the regressors - to be used by the functions \cr
\item R - total number of regressors \cr
\item num_of_models - number of models present in the model space \cr
\item forJointnes - table with model IDs and PMPs for jointness function \cr
\item forBestModels - table with model IDs, PMPs, coefficients, stds, and, stdRs for best_models function \cr
\item EMS - expected model size for binomial and binomial-beta model prior specified by the user (default EMS = R/2) \cr
\item sizePriors - table with uniform and random model priors spread over model sizes for model_sizes function \cr
\item PMPs - table with posterior model probabilities for model_sizes function \cr
\item modelPriors - table with priors on models for model_pmp function \cr
\item dilution - parameter indication if priors were diluted for model_sizes function \cr
\item alphas - coefficients on lagged dependent variable for coef_hist function\cr
\item betas_nonzero - nonzero coefficients on the regressors for coef_hist function \cr
\item d_free - table with degrees of freedom of estimated models for best_models function \cr
\item PMStable - table with prior and posterior expected model size for binomial and binomial-beta model prior
}
}
\description{
This function calculates bma object based on the model_space object obtained using optimal_model_space function.
It calculates BMA statistics and objects for the use by other functions.
}
\examples{
\donttest{
library(magrittr)

data_prepared <- bdsm::economic_growth[, 1:6] \%>\%
  bdsm::feature_standardization(
    excluded_cols = c(country, year, gdp)
  ) \%>\%
  bdsm::feature_standardization(
    group_by_col  = year,
    excluded_cols = country,
    scale         = FALSE
  )

for_bma <- bdsm::bma_prep_objects

bma_results <- bma(for_bma, df = data_prepared, round = 3, dilution = 0)
}
}
