% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bma_prep_objects}
\alias{bma_prep_objects}
\title{Example of bms_prep function result}
\format{
\subsection{\code{bma_prep_objects}}{

A list with two objects.
\describe{
\item{Table with the parameters for the entire model space}
\item{Table with likelihoods, BICs, and standard deviations, and robust standard deviations}
}
}
}
\usage{
bma_prep_objects
}
\description{
A list with two obects: model space and likelihood summary table computed using
\code{bma_prep} obtained based on the Moral-Benito data.
}
\keyword{datasets}
