% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_bma.R
\name{bma_stds}
\alias{bma_stds}
\title{Approximate standard deviations for the models}
\usage{
bma_stds(
  df,
  dep_var_col,
  timestamp_col,
  entity_col,
  model_space,
  projection_matrix_const,
  exact_value = TRUE,
  model_prior = "uniform"
)
}
\arguments{
\item{df}{Data frame with data for the SEM analysis.}

\item{dep_var_col}{Column with the dependent variable}

\item{timestamp_col}{The name of the column with timestamps}

\item{entity_col}{Column with entities (e.g. countries)}

\item{model_space}{A matrix (with named rows) with each column corresponding
to a model. Each column specifies model parameters. Compare with
\link[panels]{optimal_model_space}}

\item{projection_matrix_const}{Whether the residual maker matrix (and so
the projection matrix) should be computed for each model separately.
\code{TRUE} means that the matrix will be the same for all models}

\item{exact_value}{Whether the exact value of the likelihood should be
computed (\code{TRUE}) or just the proportional part (\code{FALSE}). Check
\link[panels]{SEM_likelihood} for details.}

\item{model_prior}{Which model prior to use. For now there are two options:
\code{'uniform'} and \code{'binomial-beta'}. Default is \code{'uniform'}.}
}
\value{
List of with two elements \code{stds} and \code{stdr}
}
\description{
Approximate standard deviations are computed for the models in the given
model space. Two versions are computed.
}
