% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{check_likelihood_data}
\alias{check_likelihood_data}
\title{Check whether the data for the likelihood function is a list
with expected named elements}
\usage{
check_likelihood_data(data)
}
\arguments{
\item{data}{argument passed to SEM_likelihood}
}
\value{
boolean
}
\description{
Check whether the data for the likelihood function is a list
with expected named elements
}
