% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{check_likelihood_params}
\alias{check_likelihood_params}
\title{Check whether the params for the likelihood function is a list
with expected named elements}
\usage{
check_likelihood_params(params)
}
\arguments{
\item{params}{argument passed to SEM_likelihood}
}
\value{
boolean
}
\description{
Check whether the params for the likelihood function is a list
with expected named elements
}
