% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_bma_params}
\alias{economic_growth_bma_params}
\title{Example Approximate Summary of Parameters of Interest Based on Model Space}
\format{
\subsection{\code{economic_growth_bma_params}}{

A double matrix with 5 rows and 8 columns
}
}
\usage{
economic_growth_bma_params
}
\description{
A matrix representing the summary of parameters computed with
\code{parameters_summary} based on the \code{economic_growth_ms} model
space. TODO: describe the matrix properly after cleaning up the code of the
function \code{parameters_summary}.
}
\keyword{datasets}
