% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_liks}
\alias{economic_growth_liks}
\title{Example Approximate Likelihoods Summary based on Model Space}
\format{
\subsection{\code{economic_growth_stds}}{

A double matrix with 11 rows and 16 columns.
\describe{
\item{first row}{Likelihoods for the models}
\item{second row}{Almost 1/2 * BIC_k as in Raftery's Bayesian Model
Selection in Social Research eq. 19.}
\item{rows 3-7}{Standard deviations}
\item{rows 8-12}{Robust standard deviations}
}
}
}
\usage{
economic_growth_liks
}
\description{
A matrix representing the summary of likelihoods computed with
\code{likelihoods_summary} based on the \code{economic_growth_ms} model
space. The matrix contains likelihoods, standard deviations and robust
standard deviations
}
\keyword{datasets}
