% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_ms}
\alias{economic_growth_ms}
\title{Example Model Space}
\format{
\subsection{\code{economic_growth_ms}}{

A double matrix with 51 rows and 16 columns.
}
}
\usage{
economic_growth_ms
}
\description{
A matrix representing the model space built using subset of regressors from
the \code{economic_growth} dataset. The included regressors are \code{ish},
\code{sed}, \code{pgrw} and \code{pop}. Therefore the model space contains
\code{2^4 = 16} models (columns).
}
\keyword{datasets}
