% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_ms_full_proj_var}
\alias{economic_growth_ms_full_proj_var}
\title{Full Model Space with Varying Projection Matrix}
\format{
\subsection{\code{economic_growth_ms_full_proj_var}}{

A double matrix with 106 rows and 512 columns.
}
}
\usage{
economic_growth_ms_full_proj_var
}
\description{
A matrix representing the model space built using all regressors from
the \code{economic_growth} dataset. Therefore the model space contains
\code{2^9 = 512} models (columns). This model space generates Posterior
Inclusion Probabilities which are consistent with the results presented by
Moral-Benito. The original results were approximated up to the 4th decimal
place. The results obtained using this model space lead to exactly the same
approximations. A different projection matrix is used for each model.
}
\keyword{datasets}
