% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_stds_robust}
\alias{economic_growth_stds_robust}
\title{Example Approximate Robust Standard Deviations based on Model Space}
\format{
\subsection{\code{economic_growth_stds_robust}}{

A double matrix with 5 rows and 16 columns.
}
}
\usage{
economic_growth_stds_robust
}
\description{
A matrix representing the robust standard deviations computed with
\code{bma_stds} based on the \code{economic_growth_ms} model space.
}
\keyword{datasets}
