% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{feature_standardization}
\alias{feature_standardization}
\title{Perform feature standarization}
\usage{
feature_standardization(
  df,
  timestamp_col,
  entity_col,
  cross_sectional = FALSE,
  scale = TRUE
)
}
\arguments{
\item{df}{Dataframe with data that should be prepared for LIML estimation}

\item{timestamp_col}{Column with timestamps (e.g. years)}

\item{entity_col}{Column with entities (e.g. countries)}

\item{cross_sectional}{Whether to perform feature standardization within
cross sections}

\item{scale}{Whether to divide by the standard deviation \code{TRUE} or not
\code{FALSE}. Default is \code{TRUE}.}
}
\description{
This function performs
\href{https://en.wikipedia.org/wiki/Feature_scaling}{feature standarization}
(also known as z-score normalization), i.e. the features are centered around
the mean and scaled with standard deviation.
}
