% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_bma_results}
\alias{full_bma_results}
\title{Example output of the bma function}
\format{
An object of class \code{list} of length 16.
}
\usage{
full_bma_results
}
\description{
A list with multiple elements summarising the BMA analysis
}
\keyword{datasets}
