% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_model_space}
\alias{full_model_space}
\title{Example of \code{model_space} function result with nine regressors (full Moral-Benito (2016) set)}
\format{
\subsection{\code{full_model_space}}{

A list with two objects.
\describe{
\item{model space}{
A double matrix with 106 rows and 512 columns with the
parameters for the model space built using all 9 regressors from
the \code{economic_growth} dataset. Therefore the model space
contains \code{2^9 = 512} models (columns).
}
\item{likelihood summmary}{
A matrix representing the summary of likelihoods computed with
\code{compute_model_space_stats} based on the model space. The first row
contains likelihoods for the models. The second row are almost
1/2 * BIC_k as in Raftery's Bayesian Model Selection in Social Research,
eq. 19. The rows 3-7 are standard deviations. Finally, the rows 8-12 are
robust standard deviations
}
}
}
}
\usage{
full_model_space
}
\description{
A list with two objects: model space and likelihood summary table computed using
\code{model_space} obtained based on the Moral-Benito (2016) data (with 9 regressors).
}
\keyword{datasets}
