% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{join_lagged_col}
\alias{join_lagged_col}
\title{Dataframe with no lagged column}
\usage{
join_lagged_col(
  df,
  col,
  col_lagged,
  timestamp_col,
  entity_col,
  timestep = NULL
)
}
\arguments{
\item{df}{Dataframe with data with a column with lagged values}

\item{col}{Column with quantity not lagged}

\item{col_lagged}{Column with the same quantity as \code{col}, but the values
are lagged in time}

\item{timestamp_col}{Column with timestamps (e.g. years)}

\item{entity_col}{Column with entities (e.g. countries)}

\item{timestep}{Difference between timestamps (e.g. 10)}
}
\value{
A dataframe with two columns merged, i.e. just one column with the desired
quantity is left.
}
\description{
This function allows to turn data in the format with lagged values for a
chosen column (i.e. there are two columns with the same quantity, but one
column is lagged in time) into the format with just one column
}
\examples{
df <- data.frame(
  entities = rep(seq(1, 4), each = 5),
  times = rep(seq(1960, 2000, 10), 4),
  value = rep(1:20),
  value_lag = lag(rep(1:20), default = 0)
)
join_lagged_col(df, value, value_lag, times, entities, timestep = 10)
}
