% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_bma.R
\name{parameters_summary}
\alias{parameters_summary}
\title{BMA summary for parameters of interest}
\usage{
parameters_summary(
  regressors,
  bet,
  pvarh,
  pvarr,
  fy,
  fyt,
  ppmsize,
  cout,
  nts,
  pts,
  variables_n
)
}
\arguments{
\item{regressors}{TODO}

\item{bet}{TODO}

\item{pvarh}{TODO}

\item{pvarr}{TODO}

\item{fy}{TODO}

\item{fyt}{TODO}

\item{ppmsize}{TODO}

\item{cout}{TODO}

\item{nts}{TODO (negatives)}

\item{pts}{TODO (positives)}

\item{variables_n}{TODO}
}
\value{
TODO dataframe with results
}
\description{
TODO This is just the code previously present in the morel-benito.R script
wrapped as a function (to get rid of the script). Well written code and docs
are still needed
}
