% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{residual_maker_matrix}
\alias{residual_maker_matrix}
\title{Residual Maker Matrix}
\usage{
residual_maker_matrix(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
M x M matrix where M is the number of rows in the \code{m} matrix.
}
\description{
Create residual maker matrix from a given matrix \code{m}. See article about
\href{https://en.wikipedia.org/wiki/Projection_matrix}{projection matrix} on
the Wikipedia.
}
