% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{safeMakeCluster}
\alias{safeMakeCluster}
\title{Version of parallel::makeCluster that checks cores limits and OS type.}
\usage{
safeMakeCluster()
}
\value{
nothing, runs makeCluster
}
\description{
Version of parallel::makeCluster that checks cores limits and OS type.
}
