% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_bma.R
\name{SEM_bma}
\alias{SEM_bma}
\title{BMA for SEM representation}
\usage{
SEM_bma(
  df,
  dep_var_col,
  timestamp_col,
  entity_col,
  projection_matrix_const,
  exact_value = TRUE,
  model_prior = "uniform",
  control = list(trace = 2, maxit = 10000, fnscale = -1, REPORT = 100)
)
}
\arguments{
\item{df}{Data frame with data for the SEM analysis.}

\item{dep_var_col}{Column with the dependent variable}

\item{timestamp_col}{The name of the column with timestamps}

\item{entity_col}{Coliumn with entities (e.g. countries)}

\item{projection_matrix_const}{Whether the residual maker matrix (and so
the projection matrix) should be computed for each model separately.
\code{TRUE} means that the matrix will be the same for all models}

\item{exact_value}{Whether the exact value of the likelihood should be
computed (\code{TRUE}) or just the proportional part (\code{FALSE}). Check
\link[panels]{SEM_likelihood} for details.}

\item{model_prior}{Which model prior to use. For now there are two options:
\code{'uniform'} and \code{'binomial-beta'}. Default is \code{'uniform'}.}

\item{control}{a list of control parameters for the optimization which are
passed to \link[stats]{optim}. Default is
\code{list(trace = 2, maxit = 10000, fnscale = -1, REPORT = 100)}, but note
that a \code{parscale} element is also added later in the function code.
For now it is hardcoded with no control on the user side.}
}
\value{
List of parameters describing analysed models
}
\description{
Perform Bayesian Model Averaging for Simultaneous Equations Model.
}
