% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{SEM_dep_var_matrix}
\alias{SEM_dep_var_matrix}
\title{Matrix with dependent variable data for SEM representation}
\usage{
SEM_dep_var_matrix(df, timestamp_col, entity_col, dep_var_col, start_time)
}
\arguments{
\item{df}{Data frame with data for the SEM analysis.}

\item{timestamp_col}{Column which determines time periods. For now only
natural numbers can be used as timestamps}

\item{entity_col}{Columns which determines entities (e.g. countires, people)}

\item{dep_var_col}{Column with dependent variable}

\item{start_time}{First time period. Only time periods greater than
\code{start_time} will be considered in the resulting matrix}
}
\value{
Matrix of size N x T where N is the number of entities considered and T is
the number of periods greater than or equal to \code{start_time}.
}
\description{
Create matrix which contains dependent variable data used in the Simultaneous
Equations Model (SEM) representation on the left hand side of the equations.
The matrix contains the data for time periods greater than chosen
\code{start_time}. The matrix is then used to compute likelihood for SEM
analysis.
}
