% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_matrices.R
\name{SEM_exogenous_matrix}
\alias{SEM_exogenous_matrix}
\title{Matrix with exogenous variables for SEM representation}
\usage{
SEM_exogenous_matrix(
  df,
  timestamp_col,
  start_time,
  lagged_col,
  regressors_subset = NULL
)
}
\arguments{
\item{df}{Data frame with data for the SEM analysis.}

\item{timestamp_col}{Column which determines time periods. For now only
natural numbers can be used as timestamps}

\item{start_time}{First time period. Only time periods greater than
\code{start_time} will be considered in the resulting matrix}

\item{lagged_col}{Column which contains lagged version of dependent variable}

\item{regressors_subset}{Which subset of columns should be used as
regressors. If \code{NULL} (default) then all remaining columns will be used
as regressors. For now columns have to be passed as list of column names
represented as strings.}
}
\value{
Matrix of size N x k+1 where N is the number of entities considered and k is
the number of chosen regressors
}
\description{
Create matrix which contains exogenous variables used in the Simultaneous
Equations Model (SEM) representation. Currently these are: dependent variable
from the \code{start_time} period and regressors from the next period after
the \code{strat_time}. The matrix is then used to compute likelihood for SEM
analysis.
}
