% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{SEM_likelihood}
\alias{SEM_likelihood}
\title{Likelihood for the SEM model}
\usage{
SEM_likelihood(
  params,
  data,
  timestamp_col = NULL,
  entity_col = NULL,
  start_time = NULL,
  lagged_col = NULL,
  dep_var_col = NULL,
  regressors = NULL,
  in_regressors = NULL,
  periods_n = NULL,
  tot_regressors_n = NULL,
  in_regressors_n = NULL,
  phis_n = NULL,
  psis_n = NULL,
  per_entity = FALSE,
  projection_matrix_const = TRUE,
  exact_value = TRUE
)
}
\arguments{
\item{params}{Parameters describing the model. Can be either a vector or a
list with named parameters.}

\item{data}{Data for the likelihood computations. Can be either a list of
matrices or a dataframe. If the dataframe, additional parameters are
required to build the matrices within the function.}

\item{timestamp_col}{Column which determines time periods. For now only
natural numbers can be used as timestampsg}

\item{entity_col}{Column which determines entities (e.g. countries, people)}

\item{start_time}{First time period. Only time periods greater than
\code{start_time} will be considered in the resulting matrix}

\item{lagged_col}{Column which contains lagged version of dependent variable}

\item{dep_var_col}{Column with dependent variable}

\item{regressors}{Which subset of columns should be used as regressors.}

\item{in_regressors}{Which subset of columns should be used as regressors
for the current model. In other words \code{regressors} are the total set of
regressors and \code{in_regressors} are the ones for which linear relation
is not set to zero for a given model.}

\item{periods_n}{Probably can be determined from the rest}

\item{tot_regressors_n}{Probably can be determined from the rest}

\item{in_regressors_n}{Probably can be determined from the rest}

\item{phis_n}{Probably can be determined from the rest}

\item{psis_n}{Probably can be determined from the rest}

\item{per_entity}{Whether to compute overall likelihood or a vector of
likelihoods with per entity value}

\item{projection_matrix_const}{Wheter the residual maker matrix (and so
the projection matrix) should be computed for each model separately.
\code{TRUE} means that the matrix will be the same for all models}

\item{exact_value}{Whether the exact value of the likelihood should be
computed (\code{TRUE}) or just the proportional part (\code{FALSE}).
Currently \code{TRUE} adds: 1. a normalization constant coming from Gaussian
distribution, 2. a term disappearing during likelihood simplification in
Likelihood-based Estimation of Dynamic Panels with Predetermined Regressors
by Moral-Benito (see Appendix A.1). The latter happens when transitioning
from equation (47) to equation (48), in step 2: the term trace{HG_22} is
dropped, because it can be assumed to be constant from Moral-Benito
perspective. To get the exact value of the likelihood we have to take this
term into account.}
}
\value{
The value of the likelihood for SEM model (or a part of interest of the
likelihood)
}
\description{
Likelihood for the SEM model
}
