% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{SEM_likelihood_bma}
\alias{SEM_likelihood_bma}
\title{Likelihood for the SEM model simplified with the assumption that
both params and data are in correct format.}
\usage{
SEM_likelihood_bma(
  params,
  data,
  timestamp_col,
  entity_col,
  dep_var_col,
  lin_related_regressors = NULL,
  per_entity = FALSE,
  projection_matrix_const = TRUE,
  exact_value = TRUE
)
}
\description{
Likelihood for the SEM model simplified with the assumption that
both params and data are in correct format.
}
\details{
Check SEM_likelihood \link[panels]{SEM_likelihood}
}
