% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_likelihood.R
\name{SEM_omega_matrix}
\alias{SEM_omega_matrix}
\title{Covariance matrix for SEM representation}
\usage{
SEM_omega_matrix(err_var, dep_vars, phis = NULL, psis = NULL)
}
\arguments{
\item{err_var}{numeric}

\item{dep_vars}{numeric vector}

\item{phis}{numeric vector}

\item{psis}{numeric vector}
}
\value{
matrix
}
\description{
Create covariance matrix for Simultaneous Equations Model (SEM)
representation. Only the part necessary to compute concentrated likelihood
function is computed (cf. Appendix in the Moral-Benito paper)
}
\examples{
err_var <- 1
dep_vars <- c(2, 2, 2, 2)
phis <- c(10, 10, 20, 20, 30, 30)
psis <- c(101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112)
SEM_omega_matrix(err_var, dep_vars, phis, psis)
}
