% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_matrices.R
\name{SEM_regressors_matrix}
\alias{SEM_regressors_matrix}
\title{Matrix with regressors data for SEM representation}
\usage{
SEM_regressors_matrix(
  df,
  timestamp_col,
  entity_col,
  regressors,
  start_time = NULL
)
}
\arguments{
\item{df}{Data frame with data for the SEM analysis.}

\item{timestamp_col}{Column which determines time periods. For now only
natural numbers can be used as timestamps}

\item{entity_col}{Column which determines entities (e.g. countries, people)}

\item{regressors}{Which subset of columns should be used as regressors.}

\item{start_time}{First time period. Only time periods greater than
\code{start_time} will be considered in the resulting matrix. For
\code{start_time=NULL} second lowest value from \code{timestamp_col} is set
as \code{start_time}. Default is \code{NULL}.}
}
\value{
Matrix of size N x (T-1)*k where N is the number of entities considered, T is
the number of periods greater than or equal to \code{start_time} and k is the
number of chosen regressors. If there are no regressors returns \code{NULL}.
}
\description{
Create matrix which contains regressors data used in the Simultaneous
Equations Model (SEM) representation on the left hand side of the equations.
The matrix contains regressors data for time periods greater than chosen
\code{start_time}. The matrix is then used to compute likelihood for SEM
analysis.
}
