% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_matrices.R
\name{sem_sigma11_matrix}
\alias{sem_sigma11_matrix}
\title{Covariance matrix for SEM representation (S11 part)}
\usage{
sem_sigma11_matrix(err_var, dep_vars)
}
\arguments{
\item{err_var}{numeric}

\item{dep_vars}{numeric vector}
}
\value{
Sigma11 matrix, its inverse and determinant
}
\description{
Create covariance matrix for Simultaneous Equations Model (SEM)
representation. Only the part necessary to compute concentrated likelihood
function is computed (cf. Appendix in the Moral-Benito paper)
}
