% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_matrices.R
\name{sem_sigma12_matrix}
\alias{sem_sigma12_matrix}
\title{Covariance matrix for SEM representation (S12 part)}
\usage{
sem_sigma12_matrix(periods_n, phis = c(), psis = c())
}
\arguments{
\item{periods_n}{numeric}

\item{phis}{numeric vector}

\item{psis}{numeric vector}
}
\value{
Sigma12 matrix
}
\description{
Create covariance matrix for Simultaneous Equations Model (SEM)
representation. Only the part necessary to compute concentrated likelihood
function is computed (cf. Appendix in the Moral-Benito paper)
}
