% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_bma.R
\name{std_dev_from_params}
\alias{std_dev_from_params}
\title{Compute Standard Deviation from Model Parameters}
\usage{
std_dev_from_params(
  params,
  df,
  timestamp_col,
  entity_col,
  dep_var_col,
  exact_value,
  projection_matrix_const,
  periods_n,
  n_entities,
  matrices_shared_across_models,
  b,
  prior_inc_prob,
  model_prior,
  regressors_n
)
}
\arguments{
\item{params}{Parameters of the model}

\item{df}{Data frame with data for the SEM analysis.}

\item{timestamp_col}{The name of the column with timestamps}

\item{entity_col}{Column with entities (e.g. countries)}

\item{dep_var_col}{Column with the dependent variable}

\item{exact_value}{Whether the exact value of the likelihood should be
computed (\code{TRUE}) or just the proportional part (\code{FALSE}). Check
\link[panels]{SEM_likelihood} for details.}

\item{projection_matrix_const}{Whether the residual maker matrix (and so
the projection matrix) should be computed for each model separately.
\code{TRUE} means that the matrix will be the same for all models}

\item{periods_n}{Number of time steps (e.g. years)}

\item{n_entities}{Number of entitied (e.g. contries)}

\item{matrices_shared_across_models}{these are Y1, Y2, Z and res_maker_matrix}

\item{b}{parameter for beta (random) distribution of the prior inclusion
probability}

\item{prior_inc_prob}{Prior inclusion probability of the model}

\item{model_prior}{Which model prior to use. For now there are two options:
\code{'uniform'} and \code{'binomial-beta'}. Default is \code{'uniform'}.}

\item{regressors_n}{Number of regressors}
}
\description{
Compute Standard Deviation from Model Parameters
}
