% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bma_prep_objects}
\alias{bma_prep_objects}
\title{Example of bma_prep function result with four regressors}
\format{
\subsection{\code{bma_prep_objects}}{

A list with two objects.
\describe{
\item{model space}{Table with the parameters for the entire model space}
\item{likelihood summmary}{Table with likelihoods, BICs, and standard deviations, and robust standard deviations}
}
}
}
\usage{
bma_prep_objects
}
\description{
A list with two objects: model space and likelihood summary table computed using
\code{bma_prep} obtained based on the Moral-Benito (2016) data (with four regressors).
}
\keyword{datasets}
