% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bma_prep_objects}
\alias{bma_prep_objects}
\title{Example of bma_prep function result with four regressors}
\format{
\subsection{\code{bma_prep_objects}}{

A list with two objects.
\describe{
\item{model space}{
A double matrix with 40 rows and 8 columns with the
parameters for the model space built using subset of the regressors from
the \code{economic_growth} dataset. The included regressors are
\code{ish}, \code{sed} and \code{pgrw}. Therefore the model space
contains \code{2^3 = 8} models (columns).
}
\item{likelihood summmary}{
A matrix representing the summary of likelihoods computed with
\code{model_space_stats} based on the model space. The first row
contains likelihoods for the models. The second row are almost
1/2 * BIC_k as in Raftery's Bayesian Model Selection in Social Research,
eq. 19. The rows 3-7 are standard deviations. Finally, the rows 8-12 are
robust standard deviations
}
}
}
}
\usage{
bma_prep_objects
}
\description{
A list with two objects: model space and likelihood summary table computed using
\code{bma_prep} obtained based on the Moral-Benito (2016) data (with four regressors).
}
\keyword{datasets}
