% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bma_prep_objects_full}
\alias{bma_prep_objects_full}
\title{Example of \code{bma_prep} function result with nine regressors (full Moral-Benito (2016) set)}
\format{
\subsection{\code{bma_prep_objects_full}}{

A list with two objects.
\describe{
\item{model space}{Table with the parameters for the entire model space}
\item{likelihood summmary}{Table with likelihoods, BICs, and standard deviations, and robust standard deviations}
}
}
}
\usage{
bma_prep_objects_full
}
\description{
A list with two objects: model space and likelihood summary table computed using
\code{bma_prep} obtained based on the Moral-Benito (2016) data (with 9 regressors).
}
\keyword{datasets}
