% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economic_growth_ms_full_proj_const}
\alias{economic_growth_ms_full_proj_const}
\title{Full Model Space with Constant Projection Matrix}
\format{
\subsection{\code{economic_growth_ms_full_proj_const}}{

A double matrix with 106 rows and 512 columns.
}
}
\usage{
economic_growth_ms_full_proj_const
}
\description{
A matrix representing the model space built using all regressors from
the \code{economic_growth} dataset. Therefore the model space contains
\code{2^9 = 512} models (columns). The same projection matrix is used for
each model.
}
\details{
TODO: to avoid NaNs when computing estimates of standard deviations, the step
size in the hessian function has to be increased to 1e-2. This is most likely
cause by the fact that the likelihood values are much closer to each other
after the correction for the projection matrix is introduced. Hence we have
to either increase the relative tolerance of the optimization algorithm or
loosen the precision when computing approximate hessian.
}
\keyword{datasets}
