% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_matrices.R
\name{SEM_B_matrix}
\alias{SEM_B_matrix}
\title{Coefficients matrix for SEM representation}
\usage{
SEM_B_matrix(alpha, periods_n, beta = c())
}
\arguments{
\item{alpha}{numeric}

\item{periods_n}{integer}

\item{beta}{numeric vector. Default is c() for no regressors case.}
}
\value{
List with two matrices B11 and B12
}
\description{
Create coefficients matrix for Simultaneous Equations Model (SEM)
representation.
}
\examples{
SEM_B_matrix(3, 4, 4:6)
}
