% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{SEM_B_matrix}
\alias{SEM_B_matrix}
\title{Coefficients matrix for SEM representation}
\usage{
SEM_B_matrix(params, regressors_n, periods_n)
}
\arguments{
\item{params}{numeric vector}

\item{regressors_n}{integer}

\item{periods_n}{integer}
}
\value{
matrix
}
\description{
Create coefficients matrix for Simultaneous Equations Model (SEM)
representation.
}
\examples{
SEM_B_matrix(3:6, 3, 4)
}
