% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_bma.R
\name{SEM_bma}
\alias{SEM_bma}
\title{BMA for SEM representation}
\usage{
SEM_bma(
  regressors_subsets,
  R_df,
  variables_n,
  regressors_n,
  periods_n,
  timestamp_col,
  year0,
  lagged_col,
  entity_col,
  Y1,
  Y2,
  res_maker_matrix,
  prandom,
  n_entities,
  b,
  pinc,
  projection_matrix_const,
  control = list(trace = 2, maxit = 10000, fnscale = -1, REPORT = 100)
)
}
\arguments{
\item{regressors_subsets}{set of regressors set. Each regressor set will be
used to construct a model.}

\item{R_df}{Data frame with data for the SEM analysis.}

\item{variables_n}{Number of variables}

\item{regressors_n}{Number of regressors}

\item{periods_n}{Number of timestamps}

\item{timestamp_col}{The name of the column with timestamps}

\item{year0}{First timestamp}

\item{lagged_col}{Column with the lagged dependent variable}

\item{entity_col}{Coliumn with entities (e.g. countries)}

\item{Y1}{Matrix with dependent variables}

\item{Y2}{Matrix with regressors}

\item{res_maker_matrix}{Residual maker matrix}

\item{prandom}{prandom = 1 for optimised_params random Ley&Steel09.
prandom = 0 for optimised_params fixed}

\item{n_entities}{Number of entities}

\item{b}{Not sure.}

\item{pinc}{Not sure. Increment for BMA}

\item{projection_matrix_const}{Wheter the residual maker matrix (and so
the projection matrix) should be computed for each model separately.
\code{TRUE} means that the matrix will be the same for all models}

\item{control}{a list of control parameters for the optimization which are
passed to \link[stats]{optim}. Default is
\code{list(trace = 2, maxit = 10000, fnscale = -1, REPORT = 100)}, but note
that a \code{parscale} element is also added later in the function code.
For now it is hardcoded with no control on the user side.}
}
\value{
List of parameters describing analysed models
}
\description{
Perform Bayesian Model Averaging for Simultaneous Equations Model.
}
